refactor: TextInput 컴포넌트 결합(#11)
